testSensors/keypad/requirementsKeypad.txt
Adafruit-Blinka==6.14.1
adafruit-circuitpython-dht==3.6.2
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
cachetools==4.2.2
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
crcmod==1.7
cryptography==2.6.1
cupshelpers==1.0
dlib==19.22.1
docutils==0.14
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.0.2
google-api-core==2.0.1
google-auth==2.1.0
google-cloud==0.34.0
google-cloud-pubsub==2.8.0
google-cloud-vision==2.6.1
googleapis-common-protos==1.53.0
gpiozero==1.6.2
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
html5lib==1.0.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
libcst==0.3.21
logilab-common==1.4.2
lxml==4.3.2
mariadb==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
mysql-connector-python==8.0.26
numpy==1.21.3
oauthlib==2.1.0
olefile==0.46
opencv-python==4.5.3.56
packaging==21.0
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
picamera==1.13
pigpio==1.78
Pillow==5.4.1
proto-plus==1.19.0
protobuf==3.18.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.53.3
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
pyusb==1.2.1
pyxdg==0.25
PyYAML==5.4.1
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-lcd==0.0.2
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
RPLCD==1.3.0
rsa==4.7.2
RTIMULib==7.2.1
schedule==1.1.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
smbus==1.1.post2
smbus2==0.4.1
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
sysv-ipc==1.1.0
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_armv7l.whl
thonny==3.3.10
twython==3.7.0
typed-ast==1.3.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
unicornhathd==0.0.4
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
